Move version command into package
